requirements.txt
# Core AI and ML Dependencies
pydantic-ai[openai]>=0.0.12
openai>=1.0.0
tiktoken>=0.5.0
langfuse>=2.0.0

# Database Dependencies
neo4j>=5.0.0
qdrant-client>=1.8.0
supabase>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0

# Web Framework Dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
httpx>=0.26.0
aiohttp>=3.9.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# Scheduling Dependencies
schedule>=1.2.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.0
jinja2>=3.1.0
starlette>=0.36.0
sse-starlette>=2.1.0

# CLI Dependencies
click>=8.1.7
rich>=13.7.0

# Configuration and Settings
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Content Processing Dependencies
beautifulsoup4>=4.12.0
markdown>=3.5.0
pypdf>=5.9.0
python-docx>=1.1.0
openpyxl>=3.1.0
nltk>=3.8.0
pytrends>=4.9.0

# Data Science Dependencies
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Google Integrations
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# Background Tasks
celery>=5.3.0

# Logging and Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# OpenTelemetry Dependencies
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
opentelemetry-exporter-prometheus>=1.12.0
opentelemetry-exporter-jaeger-thrift>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-httpx>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-psycopg2>=0.41b0
opentelemetry-instrumentation-neo4j>=0.41b0

# Utility Dependencies
tenacity>=8.2.0
cachetools>=5.3.0

# Development Dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0
pre-commit>=3.6.0
bandit>=1.7.0
safety>=2.3.0

# Additional CLI Dependencies
yaml>=6.0.0
psutil>=5.9.0

# Additional Monitoring Dependencies
aiofiles>=23.2.0
watchdog>=3.0.0
httpx>=0.25.0
